% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{clean_OTU_table}
\alias{clean_OTU_table}
\title{Clean up the OTU table}
\usage{
clean_OTU_table(OTU_file, filtered_OTU_file, OTU_table_occupancy_filter = 30)
}
\arguments{
\item{OTU_file}{pre-processed OTU table}

\item{filtered_OTU_file}{name to use for the output file (filtered OTU table)}

\item{OTU_table_occupancy_filter}{occupancy level to filter the table by}
}
\value{
None
}
\description{
Clean up the OTU table.
}
\details{
Filters out rows (OTU) by occupancy level
\itemize{
\item Takes in Data from a raw OTU table
\item Filters based on occupancy level
}
}
\note{
Original code provided output variable:  OTU_tab_sub_occ_dec
}
\examples{
# Create directory for supplementary SQL tables
#Output_dir_with_occ = 'files'
#dir.create(
#  paste0(Output_dir_with_occ, "/Supplementary/Tables_in_SQL"),
#  showWarnings = FALSE, 
#  recursive = TRUE
#)

# Clean OTU table
#clean_OTU_table(
#  'OTU_tab_file',
#  paste0(Output_dir_with_occ, "/Supplementary/Tables_in_SQL/OTU_abund.csv")
#)

}
