% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{format_otu_for_Rsqlite}
\alias{format_otu_for_Rsqlite}
\title{Put the filtered OTU, taxonomy, and abundance files into RSQLite files}
\usage{
format_otu_for_Rsqlite(
  filtered_abundance_csv,
  filtered_taxonomy_csv,
  filtered_otu_csv
)
}
\arguments{
\item{abundance_csv}{filtered OTU table}

\item{taxonomy_csv}{filtered abundance table}

\item{otu_csv}{filtered taxonomy table}
}
\value{
None
}
\description{
Put the filtered OTU, taxonomy, and abundance files into RSQLite files.
}
\details{
Reads in the filtered OTU, taxonomy, and abundance files, and creates corresponding
tables in an RSQLite file
}
\note{
format_otu_for_Rsqlite(filtered_abundance_csv, filtered_taxonomy_csv, filtered_otu_csv)
}
\examples{
# Prepare taxonomy table
#Output_dir_with_occ = 'files'
#prepair_taxonomy_table(
#  OTU_abund_filter_file = paste0(Output_dir_with_occ, "/Supplementary/Tables_in_SQL/OTU_abund.csv"),
#   input_file = 'Tax_file.csv',
#  output_file = paste0(Output_dir_with_occ, "/Supplementary/Tables_in_SQL/Taxonomy.csv")
#)

}
