% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{make_blast_py}
\alias{make_blast_py}
\title{Make BLAST}
\usage{
make_blast_py(
  filtered_OTU_file = "data/02_processed_data/filtered_otu_table.csv",
  input_fasta_file = "data/01_pre-processed_data/repseqs.fasta",
  filtered_fasta_file = "data/02_processed_data/filtered_sequences.fasta"
)
}
\description{
blast
}
