% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{maps_parallelise}
\alias{maps_parallelise}
\title{Generate maps for all OTUs in parallel}
\usage{
maps_parallelise(
  OTU_table_in = "data/02_processed_data/filtered_otu_table.csv",
  environment_data = "data/01_pre-processed_data/cs_location_avc.csv",
  Grid_file = "data/02_processed_data/ukcoast_grid.shp",
  UK_poly_file = "data/02_processed_data/ukcoast_poly.shp",
  UK_line_file = "data/02_processed_data/ukcoast_line.shp",
  maps_db = "data/18S_data/maps_db.sqlite",
  Make_png = FALSE
)
}
\arguments{
\item{OTU_table_in}{Character string. Path to the filtered OTU table.
Default is 'data/02_processed_data/filtered_otu_table.csv'.}

\item{environment_data}{Character string. Path to the environmental data file.
Default is 'data/01_pre-processed_data/cs_location_avc.csv'.}

\item{Grid_file}{Character string. Path to the interpolation grid shapefile.
Default is 'data/02_processed_data/ukcoast_grid.shp'.}

\item{UK_poly_file}{Character string. Path to the UK polygon shapefile.
Default is 'data/02_processed_data/ukcoast_poly.shp'.}

\item{UK_line_file}{Character string. Path to the UK coastline shapefile.
Default is 'data/02_processed_data/ukcoast_line.shp'.}

\item{maps_db}{Db location}

\item{Make_png}{Logical. Whether to generate PNG visualizations. Default is FALSE.}
}
\value{
None. The function creates map objects for all OTUs and stores them in the maps database.
}
\description{
Processes all OTUs in the filtered table to create map objects for each one
using parallel processing.
}
\details{
The function:
\enumerate{
\item Sets up a parallel processing cluster
\item Applies the save_otu_map function to each OTU
\item Handles environment variables for proper functioning in parallel
}
}
\note{
I think this might need merging with another function, can't remember
which or why
}
\examples{
# Prepare map
#map_prep(
#  otu_table = "/Supplementary/Tables_in_SQL/OTU_abund.csv",
#  ukcoast_poly = 'ukcoast1.shp',
#  ukcoast_line = 'ukcoast_line.shp'
#)

}
