% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{merge_AVC_location_data}
\alias{merge_AVC_location_data}
\title{Microscope: Generate taxonomic explorer app inputs}
\usage{
merge_AVC_location_data(AVC_data, CS_location_data, CS_AVC_combined)
}
\arguments{
\item{AVC_data}{Character string. Path to input file with AVC and pH data.}

\item{CS_location_data}{Character string. Path to input file with Countryside survey location data.}

\item{CS_AVC_combined}{Character string. Path where the combined output file will be saved.}
}
\value{
None. The function writes the combined data to the specified output file.
}
\description{
This script automates the process of generating a taxonomic explorer app
that links marker gene sequences to environmental responses. The application is similar to
ID-TaxER (https://shiny-apps.ceh.ac.uk/ID-TaxER/). The script creates an SQL database
and a BLAST database for the app backend, and modifies an R Shiny template file to produce
the front end. Preprocessed tables (OTU/taxonomy) and map objects are also saved locally
for reference. Currently, this script is suitable for UK data only as it uses UK map outline
objects. It should work for all CS (Countryside Survey) molecular datasets. Multiple
taxonomic datasets can be added to the same database if they share the same environmental data.

Combines AVC (Aggregate Vegetation Class) data with location information from
the Countryside survey.
}
\note{
Due to the complexity of this script, it's recommended to run it chunk by chunk
rather than knitting the entire file at once, especially during testing.
Step 0.1
Merge AVC with location from CountrySide survey

This uses the 10km areas from Countryside survey.
}
\examples{
merge_AVC_location_data("data/cs_avc_ph.csv", "data/cs_location.csv", "data/combined.csv")

}
\author{
"Briony Jones"
}
