% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_beta_pcoa}
\alias{plot_beta_pcoa}
\title{Plot PCoA with UniFrac or other distance metrics}
\usage{
plot_beta_pcoa(
  dm,
  metatable,
  grp,
  dm_name,
  permtest = FALSE,
  grp_continuous = FALSE
)
}
\arguments{
\item{dm}{A distance matrix (`dist` object or matrix).}

\item{metatable}{A data frame of metadata containing \code{sampleID} and the grouping variables.}

\item{grp}{A character string specifying the grouping variable (column) in \code{metatable}.}

\item{dm_name}{A character string for naming the distance metric (used in plot title).}

\item{permtest}{Logical; if TRUE, performs PERMANOVA (adonis2) and betadispersion tests.}

\item{grp_continuous}{Logical; if TRUE, treats the grouping variable as continuous for coloring. Default is false. It will be
automatically determined based on the property of the column}
}
\value{
A ggplot object showing the first two PCoA axes with explained variance.
}
\description{
Creates a Principal Coordinates Analysis (PCoA) plot from a distance matrix
and associated metadata, with options for PERMANOVA and PERMDISP tests and both discrete
and continuous groupings.
}
\details{
The function uses \code{\link[ape]{pcoa}} for PCoA calculation (with Cailliez correction for negative eigenvalues)
and \code{\link[ggplot2]{ggplot}} for visualization. PERMANOVA and betadispersion tests require the \pkg{vegan} package.
}
\examples{
\dontrun{
plot_unifrac(dist_mtx, metadata, "Management", "Bray-Curtis")
}

}
