% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{prepair_taxonomy_table}
\alias{prepair_taxonomy_table}
\title{Prepare taxonomy table}
\usage{
prepair_taxonomy_table(
  taxonomy_file,
  filtered_taxonomy_file,
  OTU_abund_filter_file
)
}
\arguments{
\item{taxonomy_file}{Character string. Path to the pre-processed taxonomy file.}

\item{filtered_taxonomy_file}{Character string. Path where the filtered taxonomy data will be saved.}

\item{OTU_abund_filter_file}{Character string. Path to the filtered OTU table file to use for filtering.}
}
\value{
None. The function writes the filtered taxonomy data to the specified output file.
}
\description{
Processes and formats taxonomy data for the taxonomic explorer app.
}
\details{
The function:
\enumerate{
\item Reads in the taxonomy file
\item Splits taxonomic classifications by ";" into separate columns
\item Filters to include only OTUs that meet the occupancy threshold
\item Formats taxonomy information by removing prefixes
}
}
\note{
Original code provided output variable:
Taxonomy_filt (Taxonomy_Sort is now written as extra steps were needed.)
}
\examples{
# Prepare taxonomy table
prepair_taxonomy_table(
  "data/raw_taxonomy.csv",
  "output/Supplementary/Tables_in_SQL/Taxonomy.csv",
  "output/Supplementary/Tables_in_SQL/OTU_abund.csv"
)

}
