% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{save_otu_map}
\alias{save_otu_map}
\title{Generate maps per OTU}
\usage{
save_otu_map(
  OTU_name,
  OTU_table,
  Env_table,
  Grid,
  UK_poly,
  UK_line,
  Conn,
  Schema_table_prefix,
  Output_dir,
  Make_png
)
}
\arguments{
\item{OTU_name}{Not sure what this is.}

\item{OTU_table}{filtered table}

\item{Env_table}{Env_sub #created in map_prep}

\item{Grid}{created in map_prep}

\item{UK_poly}{uk.poly #created in map_prep}

\item{UK_line}{uk.line #created in map_prep}

\item{Conn}{SQLdb, removed so correct}

\item{Schema_table_prefix}{removed for time being}

\item{Output_dir}{not used #abandoned in favour of dir in string}

\item{Make_png}{flag
#@param otu_table filtered OTU table (created in step 1.2)
#@param ukcoast_poly shape file of uk polygon
#@param ukcoast_line shape file of uk outline}
}
\value{
None
}
\description{
creates map object representing an individual.
OTUs/ ASVs geographical distribution.
}
\details{
\strong{save_otu_map} function  creates map object representing an individual
OTUs/ ASVs geographical distribution this is saved to both the specified
outputdir and database. Pngs of plotted map objects can also be saved to
outdir for reference.
}
\note{
I think this might need merging with another function, can't remember
which or why
}
\examples{

# Prepare map
#map_prep(
#  otu_table = "/Supplementary/Tables_in_SQL/OTU_abund.csv",
#  ukcoast_poly = 'ukcoast1.shp',
#  ukcoast_line = 'ukcoast_line.shp'
#)

}
