% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{clean_environmental_metadata}
\alias{clean_environmental_metadata}
\title{Microscope}
\usage{
clean_environmental_metadata(enviroment_data, filtered_env_data)
}
\arguments{
\item{enviroment_data}{enviromental data file. should contain the following column names
"avc_code","avc","pH","eastings","northings, row names should be sample ID's
(these should be able to be matched to Env file but don't need to be in the
same order)}

\item{filtered_env_data}{name to use for the outputed filtered env file}
}
\value{
None
}
\description{
Aim of script is to automate process of generating taxonomic explorer app linking marker gene sequences to environmental responses (see ID-TaxER https://shiny-apps.ceh.ac.uk/ID-TaxER/ for example of similar existing app). Script directly adds to SQL database and generates blast database for app backend and modifies r shiny template file to produce front end. Preprocessed tables (otu/taxonomy etc) and map objects are also saved locally for reference. At the moment this script is suitable for UK data only as uses UK map outline objects to run, should be suitable for all CS molecular datasets (which was really my motivation for writing this). Due to complexity of script would recommend running code chunk by code chunk, rather than knitting it (atleast while it is still being tested). Apps for multiple taxonomic datasets can be generated using repeated runs of this script and added to the same database if they share the same environmental data.

#@param OTU_tab_file should have OTU's/ASV's as column names and sample IDs as row
#@param Tax_file with two columns one with OTU/ASV names and the other with
#@param Fasta_file with OTU representative sequences/ ASV sequences corresponding to
#@param Env_file should contain the following column names
#@param App_template_input_dir location of shiny app template

#@param OTU_tab_occ_filter: The minimum amount of samples an OTU should occur in to be included in app
#@param Map_objs_input_dir: Directory with input map objects files
#@param Output_dir: Directory for all outputs
#@param SQL_database_name: Name of database to input pre-processed tables and maps,
#@param SQL_database_host: Database host address for data inputs
#@param Schema_table_prefix: This string will be added to database schema and table
#@param Empty_database: TRUE/FALSE, if database is empty all database schemas will be
#@param Use_occupancy_in_schema_table_names: TRUE/FALSE, if TRUE will add occupancy
#@param App_title: Title to appear on the app header (not linked to future web
#@param Example_sequence: Sequence to be used as an example sequence for users when
#@param Info_text: Text to feature on app describing purpose and function

Clean up enviromental metadata

Clean up enviromental metadata.
}
\details{
Generate taxonomic explorer app inputs

Takes in pre-prcoessed data and cleans it up:
\itemize{
\item Removes unneeded fields
\item rearranges into a format SQL is happier with
\item Write new data to file
}
}
\note{
Original code provided output variable: Env_for_SQL
}
\examples{
# Create directory for supplementary SQL tables
#dir.create(
#  'Supplementary/Tables_in_SQL,
#  showWarnings = FALSE, 
#  recursive = TRUE
#)

# Clean environmental metadata
#clean_environmental_metadata(
#  params$Env_file,
#  paste0(Output_dir_with_occ, "/Supplementary/Tables_in_SQL/Env.csv")
#)

}
\author{
"Briony Jones"
}
