% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{clean_environmental_metadata}
\alias{clean_environmental_metadata}
\title{Clean up enviromental metadata}
\usage{
clean_environmental_metadata(enviroment_data, filtered_env_data)
}
\arguments{
\item{enviroment_data}{enviromental data file. should contain the following column names
"avc_code","avc","pH","eastings","northings, row names should be sample ID's
(these should be able to be matched to Env file but don't need to be in the
same order)}

\item{filtered_env_data}{name to use for the outputed filtered env file}
}
\value{
None
}
\description{
Clean up enviromental metadata.
}
\details{
Takes in pre-prcoessed data and cleans it up:
\itemize{
\item Removes unneeded fields
\item rearranges into a format SQL is happier with
\item Write new data to file
}
}
\note{
Original code provided output variable: Env_for_SQL
}
\examples{
# Create directory for supplementary SQL tables
#dir.create(
#  'Supplementary/Tables_in_SQL,
#  showWarnings = FALSE, 
#  recursive = TRUE
#)

# Clean environmental metadata
#clean_environmental_metadata(
#  params$Env_file,
#  paste0(Output_dir_with_occ, "/Supplementary/Tables_in_SQL/Env.csv")
#)

}
