% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_abundance_stats}
\alias{get_abundance_stats}
\title{Prepare abundance_stats table}
\usage{
get_abundance_stats(filtered_OTU_file, abundance_stats_file)
}
\arguments{
\item{filtered_OTU_file}{filtered OTU table (created in step 1.2)}

\item{abundance_stats_file}{name to use for the output file, filtered abundace stats}
}
\value{
None
}
\description{
Prepare abundance_stats table.
}
\details{
Get individual OTU stats to summarise abundance (rank) and occupancy
(percentage and rank), these stats will form a table in the database (added to
DB in step 2). In this code chunk saved to r
params$Output_dir /Supplementary/Tables_in_SQL subfolder for reference.
}
\note{
Original code provided output variable:  abundance_stats
}
\examples{
# Requires OTU_tab_sub_occ_dec variable created in another function
# Get abundance statistics
#Output_dir_with_occ = 'files'
#get_abundance_stats(
#  paste0(Output_dir_with_occ, "/Supplementary/Tables_in_SQL/OTU_abund.csv"),
#  paste0(Output_dir_with_occ, "/Supplementary/Tables_in_SQL/abundance_stats.csv")
#)

}
