% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{map_prep}
\alias{map_prep}
\title{Put the filtered OTU, taxonomy, and abundance files into RSQLite files}
\usage{
map_prep(
  ukcoast_poly,
  ukcoast_line,
  uk_poly_converted,
  uk_line_converted,
  uk_grid_converted
)
}
\arguments{
\item{ukcoast_poly}{shape file of uk polygon}

\item{ukcoast_line}{shape file of uk outline}

\item{uk_poly_converted}{shape file of uk polygon}

\item{uk_line_converted}{shape file of uk outline}

\item{uk_grid_converted}{shape file of uk outline}
}
\value{
None
}
\description{
Step 3 Make map objects for DB.
}
\details{
Step 3 Make map objects for DB
}
\note{
Running this as separate function, but perhaps this should be the start of another function?

abund_table is called by env name: OTU_tab_sub_occ_dec
looks like this sets up a series of functions and variables to be used in a latter step
to make a db. This last part maybe much harder to split up.   Need to focus on first
parts and leave this until later. Perhaps it would be better to move to SQLite????
Perhaps some of the functions to setup, like grid bit
}
\examples{
# Prepare map
#map_prep(
#  otu_table = "supplementary/Tables_in_SQL/OTU_abund.csv",
#  enviroment_data='country_side_survey/CS2007_Env_pH_AVC.csv'
#  ukcoast_poly = 'ukcoast1.shp',
#  ukcoast_line = 'ukcoast_line.shp'
#)

}
