% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{maps_parallelise}
\alias{maps_parallelise}
\title{Generate maps per OTU}
\usage{
maps_parallelise(
  Output_dir_with_occ,
  OTU_tab_sub_occ_dec,
  Env_sub,
  grd,
  uk.poly,
  uk.line,
  Schema_table_prefix_modified
)
}
\arguments{
\item{Output_dir_with_occ}{blank}

\item{OTU_tab_sub_occ_dec}{blank}

\item{Env_sub}{blank}

\item{grd}{blank}

\item{uk.poly}{blank}

\item{uk.line}{blank}

\item{Schema_table_prefix_modified}{blank}
}
\value{
None
}
\description{
Using rparallel to parallelise -need to reauthenticate database on all cpus
working on.
}
\details{
Runs save_otu_map function on all OTUs.
Using rparallel to parallelise -need to reauthenticate database on all cpus
working on.
}
\note{
I think this might need merging with another function, can't remember
which or why
}
\examples{
# Prepare map
#map_prep(
#  otu_table = "/Supplementary/Tables_in_SQL/OTU_abund.csv",
#  ukcoast_poly = 'ukcoast1.shp',
#  ukcoast_line = 'ukcoast_line.shp'
#)

}
