% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{prepair_taxonomy_table}
\alias{prepair_taxonomy_table}
\title{Prepare taxonomy table}
\usage{
prepair_taxonomy_table(
  taxonomy_file,
  filtered_taxonomy_file,
  OTU_abund_filter_file
)
}
\arguments{
\item{taxonomy_file}{pre-processed taxonomy table to use as input}

\item{filtered_taxonomy_file}{name for output file, which is a filtered taxonomy table}

\item{OTU_abund_filter_file}{filtered OTU table created in step 1.2 for input}
}
\value{
None
}
\description{
Prepare abundance_stats table.
}
\details{
Read in taxonomy, split taxonomic fields by ";"and subset to only include OTUs
that meet occupancy threshold ,save to r params$Output_dir/Supplementary/Tables_in_SQL
subfolder for reference.
}
\note{
Original code provided output variable:
Taxonomy_filt (Taxonomy_Sort is now written as extra steps were needed.)
}
\examples{
# Prepare taxonomy table
#Output_dir_with_occ = 'files'
#prepair_taxonomy_table(
#  OTU_abund_filter_file = paste0(Output_dir_with_occ, "/Supplementary/Tables_in_SQL/OTU_abund.csv")#,
 # input_file = 'Tax_file',
 # output_file = paste0(Output_dir_with_occ, "/Supplementary/Tables_in_SQL/Taxonomy.csv")
#)

}
